[add] UrlNotFoundException [refactor] 코딩 컨벤션, import 정리
