--- src/main/java/com/pppp0722/springbooturlshortener/exception/IllegalUrlArgumentException.java
@@ -1,11 +1,0 @@
-package com.pppp0722.springbooturlshortener.exception;
-
-public class IllegalUrlArgumentException extends RuntimeException {
-
-    public IllegalUrlArgumentException() {
-    }
-
-    public IllegalUrlArgumentException(String message) {
-        super(message);
-    }
-}
