+++ src/main/java/com/pppp0722/springbooturlshortener/exception/UrlIllegalArgumentException.java
@@ -1,0 +1,11 @@
+package com.pppp0722.springbooturlshortener.exception;
+
+public class UrlIllegalArgumentException extends RuntimeException {
+
+    public UrlIllegalArgumentException() {
+    }
+
+    public UrlIllegalArgumentException(String message) {
+        super(message);
+    }
+}
