+++ src/main/java/com/pppp0722/springbooturlshortener/exception/UrlNotFoundException.java
@@ -1,0 +1,11 @@
+package com.pppp0722.springbooturlshortener.exception;
+
+public class UrlNotFoundException extends RuntimeException{
+
+    public UrlNotFoundException() {
+    }
+
+    public UrlNotFoundException(String message) {
+        super(message);
+    }
+}
