+++ src/main/resources/templates/error/400.html
@@ -1,0 +1,19 @@
+<!doctype html>
+<html lang="en">
+<head>
+  <title>URL 단축 서비스</title>
+  <style>
+  </style>
+</head>
+<body>
+<div>
+  <a th:href="@{/}">메인 페이지로 이동</a>
+</div>
+<div>
+  400 에러
+</div>
+<div>
+  message: <span th:text="${message}"/>
+</div>
+</body>
+</html>
