+++ src/main/resources/templates/error/404.html
@@ -10,7 +10,7 @@
   <a th:href="@{/}">메인 페이지로 이동</a>
 </div>
 <div>
-  401 에러
+  404 에러
 </div>
 <div>
   message: <span th:text="${message}"/>
